Convert autoupdate_* resources to the new tctl handler format
