+++ tool/tctl/common/edit_command_test.go
@@ -597,7 +597,7 @@ func testEditAutoUpdateConfig(t *testing.T, clt *authclient.Client) {
 			return trace.Wrap(err, "opening file to edit")
 		}
 		expected.GetMetadata().Revision = initial.GetMetadata().GetRevision()
-		collection := &autoUpdateConfigCollection{config: expected}
+		collection := resources.NewAutoUpdateConfigCollection(expected)
 		return trace.NewAggregate(writeYAML(collection, f), f.Close())
 	}
 
@@ -639,7 +639,7 @@ func testEditAutoUpdateVersion(t *testing.T, clt *authclient.Client) {
 			return trace.Wrap(err, "opening file to edit")
 		}
 		expected.GetMetadata().Revision = initial.GetMetadata().GetRevision()
-		collection := &autoUpdateVersionCollection{version: expected}
+		collection := resources.NewAutoUpdateVersionCollection(expected)
 		return trace.NewAggregate(writeYAML(collection, f), f.Close())
 	}
