+++ tool/tctl/common/resource_command.go
@@ -41,13 +41,11 @@ import (
 	"github.com/gravitational/teleport/api/client/proto"
 	apidefaults "github.com/gravitational/teleport/api/defaults"
 	accessmonitoringrulesv1pb "github.com/gravitational/teleport/api/gen/proto/go/teleport/accessmonitoringrules/v1"
-	autoupdatev1pb "github.com/gravitational/teleport/api/gen/proto/go/teleport/autoupdate/v1"
 	clusterconfigpb "github.com/gravitational/teleport/api/gen/proto/go/teleport/clusterconfig/v1"
 	crownjewelv1 "github.com/gravitational/teleport/api/gen/proto/go/teleport/crownjewel/v1"
 	dbobjectv1 "github.com/gravitational/teleport/api/gen/proto/go/teleport/dbobject/v1"
 	dbobjectimportrulev1 "github.com/gravitational/teleport/api/gen/proto/go/teleport/dbobjectimportrule/v1"
 	devicepb "github.com/gravitational/teleport/api/gen/proto/go/teleport/devicetrust/v1"
-	headerv1 "github.com/gravitational/teleport/api/gen/proto/go/teleport/header/v1"
 	healthcheckconfigv1 "github.com/gravitational/teleport/api/gen/proto/go/teleport/healthcheckconfig/v1"
 	loginrulepb "github.com/gravitational/teleport/api/gen/proto/go/teleport/loginrule/v1"
 	pluginsv1 "github.com/gravitational/teleport/api/gen/proto/go/teleport/plugins/v1"
@@ -166,10 +164,6 @@ func (rc *ResourceCommand) Initialize(app *kingpin.Application, _ *tctlcfg.Globa
 		types.KindPlugin:                      rc.createPlugin,
 		types.KindStaticHostUser:              rc.createStaticHostUser,
 		types.KindUserTask:                    rc.createUserTask,
-		types.KindAutoUpdateConfig:            rc.createAutoUpdateConfig,
-		types.KindAutoUpdateVersion:           rc.createAutoUpdateVersion,
-		types.KindAutoUpdateAgentRollout:      rc.createAutoUpdateAgentRollout,
-		types.KindAutoUpdateAgentReport:       rc.upsertAutoUpdateAgentReport,
 		types.KindHealthCheckConfig:           rc.createHealthCheckConfig,
 		scopedaccess.KindScopedRole:           rc.createScopedRole,
 		scopedaccess.KindScopedRoleAssignment: rc.createScopedRoleAssignment,
@@ -191,11 +185,7 @@ func (rc *ResourceCommand) Initialize(app *kingpin.Application, _ *tctlcfg.Globa
 		types.KindPlugin:                      rc.updatePlugin,
 		types.KindStaticHostUser:              rc.updateStaticHostUser,
 		types.KindUserTask:                    rc.updateUserTask,
-		types.KindAutoUpdateConfig:            rc.updateAutoUpdateConfig,
-		types.KindAutoUpdateVersion:           rc.updateAutoUpdateVersion,
 		types.KindDynamicWindowsDesktop:       rc.updateDynamicWindowsDesktop,
-		types.KindAutoUpdateAgentRollout:      rc.updateAutoUpdateAgentRollout,
-		types.KindAutoUpdateAgentReport:       rc.upsertAutoUpdateAgentReport,
 		types.KindHealthCheckConfig:           rc.updateHealthCheckConfig,
 		scopedaccess.KindScopedRole:           rc.updateScopedRole,
 		scopedaccess.KindScopedRoleAssignment: rc.updateScopedRoleAssignment,
@@ -1413,9 +1403,6 @@ func (rc *ResourceCommand) Delete(ctx context.Context, client *authclient.Client
 		types.KindInstaller,
 		types.KindUIConfig,
 		types.KindNetworkRestrictions,
-		types.KindAutoUpdateConfig,
-		types.KindAutoUpdateVersion,
-		types.KindAutoUpdateAgentRollout,
 	}
 	if !slices.Contains(singletonResources, rc.ref.Kind) && (rc.ref.Kind == "" || rc.ref.Name == "") {
 		return trace.BadParameter("provide a full resource name to delete, for example:\n$ tctl rm cluster/east\n")
@@ -1750,21 +1737,6 @@ func (rc *ResourceCommand) Delete(ctx context.Context, client *authclient.Client
 			return trace.Wrap(err)
 		}
 		fmt.Printf("static host user %q has been deleted\n", rc.ref.Name)
-	case types.KindAutoUpdateConfig:
-		if err := client.DeleteAutoUpdateConfig(ctx); err != nil {
-			return trace.Wrap(err)
-		}
-		fmt.Printf("AutoUpdateConfig has been deleted\n")
-	case types.KindAutoUpdateVersion:
-		if err := client.DeleteAutoUpdateVersion(ctx); err != nil {
-			return trace.Wrap(err)
-		}
-		fmt.Printf("AutoUpdateVersion has been deleted\n")
-	case types.KindAutoUpdateAgentRollout:
-		if err := client.DeleteAutoUpdateAgentRollout(ctx); err != nil {
-			return trace.Wrap(err)
-		}
-		fmt.Printf("AutoUpdateAgentRollout has been deleted\n")
 	case types.KindHealthCheckConfig:
 		return trace.Wrap(rc.deleteHealthCheckConfig(ctx, client))
 	case types.KindRelayServer:
@@ -2601,38 +2573,6 @@ func (rc *ResourceCommand) getCollection(ctx context.Context, client *authclient
 			return nil, trace.Wrap(err)
 		}
 		return &staticHostUserCollection{items: resources}, nil
-	case types.KindAutoUpdateConfig:
-		config, err := client.GetAutoUpdateConfig(ctx)
-		if err != nil {
-			return nil, trace.Wrap(err)
-		}
-		return &autoUpdateConfigCollection{config}, nil
-	case types.KindAutoUpdateVersion:
-		version, err := client.GetAutoUpdateVersion(ctx)
-		if err != nil {
-			return nil, trace.Wrap(err)
-		}
-		return &autoUpdateVersionCollection{version}, nil
-	case types.KindAutoUpdateAgentRollout:
-		version, err := client.GetAutoUpdateAgentRollout(ctx)
-		if err != nil {
-			return nil, trace.Wrap(err)
-		}
-		return &autoUpdateAgentRolloutCollection{version}, nil
-	case types.KindAutoUpdateAgentReport:
-		if rc.ref.Name != "" {
-			report, err := client.GetAutoUpdateAgentReport(ctx, rc.ref.Name)
-			if err != nil {
-				return nil, trace.Wrap(err)
-			}
-			return &autoUpdateAgentReportCollection{reports: []*autoupdatev1pb.AutoUpdateAgentReport{report}}, nil
-		}
-
-		reports, err := stream.Collect(clientutils.Resources(ctx, client.ListAutoUpdateAgentReports))
-		if err != nil {
-			return nil, trace.Wrap(err)
-		}
-		return &autoUpdateAgentReportCollection{reports: reports}, nil
 	case types.KindAccessMonitoringRule:
 		if rc.ref.Name != "" {
 			rule, err := client.AccessMonitoringRuleClient().GetAccessMonitoringRule(ctx, rc.ref.Name)
@@ -3005,158 +2945,5 @@ func (rc *ResourceCommand) updateAccessGraphSettings(ctx context.Context, client
 		return trace.Wrap(err)
 	}
 	fmt.Println("access_graph_settings has been updated")
-	return nil
-}
-
-func (rc *ResourceCommand) createAutoUpdateConfig(ctx context.Context, client *authclient.Client, raw services.UnknownResource) error {
-	config, err := services.UnmarshalProtoResource[*autoupdatev1pb.AutoUpdateConfig](raw.Raw, services.DisallowUnknown())
-	if err != nil {
-		return trace.Wrap(err)
-	}
-
-	if config.GetMetadata() == nil {
-		config.Metadata = &headerv1.Metadata{}
-	}
-	if config.GetMetadata().GetName() == "" {
-		config.Metadata.Name = types.MetaNameAutoUpdateConfig
-	}
-
-	if rc.IsForced() {
-		_, err = client.UpsertAutoUpdateConfig(ctx, config)
-	} else {
-		_, err = client.CreateAutoUpdateConfig(ctx, config)
-	}
-	if err != nil {
-		return trace.Wrap(err)
-	}
-
-	fmt.Println("autoupdate_config has been created")
-	return nil
-}
-
-func (rc *ResourceCommand) updateAutoUpdateConfig(ctx context.Context, client *authclient.Client, raw services.UnknownResource) error {
-	config, err := services.UnmarshalProtoResource[*autoupdatev1pb.AutoUpdateConfig](raw.Raw, services.DisallowUnknown())
-	if err != nil {
-		return trace.Wrap(err)
-	}
-
-	if config.GetMetadata() == nil {
-		config.Metadata = &headerv1.Metadata{}
-	}
-	if config.GetMetadata().GetName() == "" {
-		config.Metadata.Name = types.MetaNameAutoUpdateConfig
-	}
-
-	if _, err := client.UpdateAutoUpdateConfig(ctx, config); err != nil {
-		return trace.Wrap(err)
-	}
-	fmt.Println("autoupdate_config has been updated")
-	return nil
-}
-
-func (rc *ResourceCommand) createAutoUpdateVersion(ctx context.Context, client *authclient.Client, raw services.UnknownResource) error {
-	version, err := services.UnmarshalProtoResource[*autoupdatev1pb.AutoUpdateVersion](raw.Raw, services.DisallowUnknown())
-	if err != nil {
-		return trace.Wrap(err)
-	}
-
-	if version.GetMetadata() == nil {
-		version.Metadata = &headerv1.Metadata{}
-	}
-	if version.GetMetadata().GetName() == "" {
-		version.Metadata.Name = types.MetaNameAutoUpdateVersion
-	}
-
-	if rc.IsForced() {
-		_, err = client.UpsertAutoUpdateVersion(ctx, version)
-	} else {
-		_, err = client.CreateAutoUpdateVersion(ctx, version)
-	}
-	if err != nil {
-		return trace.Wrap(err)
-	}
-
-	fmt.Println("autoupdate_version has been created")
-	return nil
-}
-
-func (rc *ResourceCommand) updateAutoUpdateVersion(ctx context.Context, client *authclient.Client, raw services.UnknownResource) error {
-	version, err := services.UnmarshalProtoResource[*autoupdatev1pb.AutoUpdateVersion](raw.Raw, services.DisallowUnknown())
-	if err != nil {
-		return trace.Wrap(err)
-	}
-
-	if version.GetMetadata() == nil {
-		version.Metadata = &headerv1.Metadata{}
-	}
-	if version.GetMetadata().GetName() == "" {
-		version.Metadata.Name = types.MetaNameAutoUpdateVersion
-	}
-
-	if _, err := client.UpdateAutoUpdateVersion(ctx, version); err != nil {
-		return trace.Wrap(err)
-	}
-	fmt.Println("autoupdate_version has been updated")
-	return nil
-}
-
-func (rc *ResourceCommand) createAutoUpdateAgentRollout(ctx context.Context, client *authclient.Client, raw services.UnknownResource) error {
-	rollout, err := services.UnmarshalProtoResource[*autoupdatev1pb.AutoUpdateAgentRollout](raw.Raw, services.DisallowUnknown())
-	if err != nil {
-		return trace.Wrap(err)
-	}
-
-	if rollout.GetMetadata() == nil {
-		rollout.Metadata = &headerv1.Metadata{}
-	}
-	if rollout.GetMetadata().GetName() == "" {
-		rollout.Metadata.Name = types.MetaNameAutoUpdateAgentRollout
-	}
-
-	if rc.IsForced() {
-		_, err = client.UpsertAutoUpdateAgentRollout(ctx, rollout)
-	} else {
-		_, err = client.CreateAutoUpdateAgentRollout(ctx, rollout)
-	}
-	if err != nil {
-		return trace.Wrap(err)
-	}
-
-	fmt.Println("autoupdate_agent_rollout has been created")
-	return nil
-}
-
-func (rc *ResourceCommand) upsertAutoUpdateAgentReport(ctx context.Context, client *authclient.Client, raw services.UnknownResource) error {
-	report, err := services.UnmarshalProtoResource[*autoupdatev1pb.AutoUpdateAgentReport](raw.Raw, services.DisallowUnknown())
-	if err != nil {
-		return trace.Wrap(err)
-	}
-
-	_, err = client.UpsertAutoUpdateAgentReport(ctx, report)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-
-	fmt.Println("autoupdate_agent_report has been created")
-	return nil
-}
-
-func (rc *ResourceCommand) updateAutoUpdateAgentRollout(ctx context.Context, client *authclient.Client, raw services.UnknownResource) error {
-	rollout, err := services.UnmarshalProtoResource[*autoupdatev1pb.AutoUpdateAgentRollout](raw.Raw, services.DisallowUnknown())
-	if err != nil {
-		return trace.Wrap(err)
-	}
-
-	if rollout.GetMetadata() == nil {
-		rollout.Metadata = &headerv1.Metadata{}
-	}
-	if rollout.GetMetadata().GetName() == "" {
-		rollout.Metadata.Name = types.MetaNameAutoUpdateAgentRollout
-	}
-
-	if _, err := client.UpdateAutoUpdateAgentRollout(ctx, rollout); err != nil {
-		return trace.Wrap(err)
-	}
-	fmt.Println("autoupdate_version has been updated")
 	return nil
 }
