+++ tool/tctl/common/resources/resource.go
@@ -37,7 +37,11 @@ func Handlers() map[string]Handler {
 		types.KindApp:                                appHandler(),
 		types.KindAppServer:                          appServerHandler(),
 		types.KindAuthServer:                         authHandler(),
+		types.KindAutoUpdateAgentReport:              autoUpdateAgentReportHandler(),
+		types.KindAutoUpdateAgentRollout:             autoUpdateAgentRolloutHandler(),
 		types.KindAutoUpdateBotInstanceReport:        autoUpdateBotInstanceReportHandler(),
+		types.KindAutoUpdateConfig:                   autoUpdateConfigHandler(),
+		types.KindAutoUpdateVersion:                  autoUpdateVersionHandler(),
 		types.KindBot:                                botHandler(),
 		types.KindBotInstance:                        botInstanceHandler(),
 		types.KindDatabase:                           databaseHandler(),
